Move side effect of queryChannel to plugin
