+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/channel/ChannelController.kt
@@ -72,22 +72,17 @@ import io.getstream.chat.android.client.utils.map
 import io.getstream.chat.android.client.utils.recover
 import io.getstream.chat.android.core.ExperimentalStreamChatApi
 import io.getstream.chat.android.offline.ChatDomainImpl
+import io.getstream.chat.android.offline.experimental.channel.logic.ChannelLogic
 import io.getstream.chat.android.offline.experimental.channel.state.ChannelMutableState
-import io.getstream.chat.android.offline.experimental.querychannels.logic.QueryChannelLogic
 import io.getstream.chat.android.offline.extensions.addMyReaction
-import io.getstream.chat.android.offline.extensions.inOffsetWith
 import io.getstream.chat.android.offline.extensions.isPermanent
 import io.getstream.chat.android.offline.extensions.removeMyReaction
 import io.getstream.chat.android.offline.message.MessageSendingService
 import io.getstream.chat.android.offline.message.MessageSendingServiceFactory
-import io.getstream.chat.android.offline.message.NEVER
 import io.getstream.chat.android.offline.message.attachment.AttachmentUploader
-import io.getstream.chat.android.offline.message.attachment.AttachmentUrlValidator
 import io.getstream.chat.android.offline.message.isEphemeral
-import io.getstream.chat.android.offline.message.shouldIncrementUnreadCount
 import io.getstream.chat.android.offline.message.wasCreatedAfter
 import io.getstream.chat.android.offline.message.wasCreatedBeforeOrAt
-import io.getstream.chat.android.offline.model.ChannelConfig
 import io.getstream.chat.android.offline.request.QueryChannelPaginationRequest
 import io.getstream.chat.android.offline.thread.ThreadController
 import kotlinx.coroutines.Job
@@ -102,16 +97,14 @@ import java.io.File
 import java.util.Calendar
 import java.util.Date
 import java.util.concurrent.ConcurrentHashMap
-import kotlin.math.max
 
 @OptIn(ExperimentalStreamChatApi::class)
 public class ChannelController internal constructor(
     private val mutableState: ChannelMutableState,
-    private val queryChannelLogic: QueryChannelLogic,
+    private val channelLogic: ChannelLogic,
     private val client: ChatClient,
     @VisibleForTesting
     internal val domainImpl: ChatDomainImpl,
-    private val attachmentUrlValidator: AttachmentUrlValidator = AttachmentUrlValidator(),
     private val attachmentUploader: AttachmentUploader = AttachmentUploader(client),
     messageSendingServiceFactory: MessageSendingServiceFactory = MessageSendingServiceFactory(),
 ) {
@@ -270,7 +263,7 @@ public class ChannelController internal constructor(
     }
 
     internal suspend fun hide(clearHistory: Boolean): Result<Unit> {
-        setHidden(true)
+        channelLogic.setHidden(true)
         val result = channelClient.hide(clearHistory).await()
         if (result.isSuccess) {
             if (clearHistory) {
@@ -287,7 +280,7 @@ public class ChannelController internal constructor(
     }
 
     internal suspend fun show(): Result<Unit> {
-        setHidden(false)
+        channelLogic.setHidden(false)
         val result = channelClient.show().await()
         if (result.isSuccess) {
             domainImpl.repos.setHiddenForChannel(cid, false)
@@ -411,70 +404,22 @@ public class ChannelController internal constructor(
     }
 
     private suspend fun runChannelQuery(request: WatchChannelRequest): Result<Channel> {
-        val preconditionResult = queryChannelLogic.onQueryChannelPrecondition(channelType, channelId, request)
+        val preconditionResult = channelLogic.onQueryChannelPrecondition(channelType, channelId, request)
         if (preconditionResult.isError) {
             return Result.error(preconditionResult.error())
         }
-        val loader = queryChannelLogic.loadingStateByRequest(request)
-        loader.value = true
-        // first we load the data from room and update the messages and channel flow
-        val queryOfflineJob = domainImpl.scope.async { runChannelQueryOffline(request) }
 
-        // start the online query before queryOfflineJob.await
-        val queryOnlineJob = domainImpl.scope.async { runChannelQueryOnline(request) }
-        val localChannel = queryOfflineJob.await()?.also { channel ->
-            if (request.filteringOlderMessages()) {
-                updateOldMessagesFromLocalChannel(channel)
-            } else {
-                updateDataFromLocalChannel(channel)
-            }
-            loader.value = false
+        val offlineChannel = channelLogic.runChannelQueryOffline(request)
+
+        val onlineResult = client.queryChannelInternal(channelType, channelId, request).await().also { result ->
+            channelLogic.onQueryChannelResult(result, channelType, channelId, request)
         }
 
-        val result: Result<Channel> = queryOnlineJob.await().let { onlineResult ->
-            if (onlineResult.isSuccess) {
-                onlineResult.also { updateDataFromChannel(onlineResult.data()) }
-            } else {
-                if (onlineResult.error().isPermanent()) {
-                    logger.logW("Permanent failure calling channel.watch for channel $cid, with error ${onlineResult.error()}")
-                } else {
-                    logger.logW("Temporary failure calling channel.watch for channel $cid. Marking the channel as needing recovery. Error was ${onlineResult.error()}")
-                    mutableState.recoveryNeeded = true
-                }
-                localChannel?.let { Result(it) } ?: onlineResult
-            }
+        return when {
+            onlineResult.isSuccess -> onlineResult
+            offlineChannel != null -> Result.success(offlineChannel)
+            else -> onlineResult
         }
-        loader.value = false
-        return result
-    }
-
-    private suspend fun runChannelQueryOffline(pagination: WatchChannelRequest): Channel? =
-        domainImpl.selectAndEnrichChannel(cid, pagination)?.also { channel ->
-            logger.logI("Loaded channel ${channel.cid} from offline storage with ${channel.messages.size} messages")
-        }
-
-    private suspend fun runChannelQueryOnline(request: WatchChannelRequest): Result<Channel> {
-        val response = channelClient.watch(request).await()
-
-        if (response.isSuccess) {
-            mutableState.recoveryNeeded = false
-            val channelResponse = response.data()
-            if (request.messagesLimit() > channelResponse.messages.size) {
-                if (request.isFilteringNewerMessages()) {
-                    mutableState._endOfNewerMessages.value = true
-                } else {
-                    mutableState._endOfOlderMessages.value = true
-                }
-            }
-            // first thing here needs to be updating configs otherwise we have a race with receiving events
-            domainImpl.repos.insertChannelConfig(ChannelConfig(channelResponse.type, channelResponse.config))
-
-            domainImpl.storeStateForChannel(channelResponse)
-        } else {
-            mutableState.recoveryNeeded = true
-            domainImpl.addError(response.error())
-        }
-        return response
     }
 
     internal suspend fun sendMessage(message: Message): Result<Message> = messageSendingService.sendNewMessage(message)
@@ -729,16 +674,10 @@ public class ChannelController internal constructor(
         }
     }
 
-    private fun setWatcherCount(watcherCount: Int) {
-        if (watcherCount != mutableState._watcherCount.value) {
-            mutableState._watcherCount.value = watcherCount
-        }
-    }
-
     // This one needs to be public for flows such as running a message action
 
     internal fun upsertMessage(message: Message) {
-        upsertMessages(listOf(message))
+        channelLogic.upsertMessages(listOf(message))
     }
 
     private fun upsertEventMessage(message: Message) {
@@ -746,7 +685,7 @@ public class ChannelController internal constructor(
         getMessage(message.id)?.let {
             message.ownReactions = it.ownReactions
         }
-        upsertMessages(listOf(message))
+        channelLogic.upsertMessages(listOf(message))
     }
 
     public fun getMessage(messageId: String): Message? {
@@ -760,43 +699,6 @@ public class ChannelController internal constructor(
         }
 
         return message
-    }
-
-    private fun parseMessages(messages: List<Message>): Map<String, Message> {
-        val currentMessages = mutableState._messages.value
-        return currentMessages + attachmentUrlValidator.updateValidAttachmentsUrl(messages, currentMessages)
-            .filter { newMessage -> isMessageNewerThanCurrent(currentMessages[newMessage.id], newMessage) }
-            .associateBy(Message::id)
-    }
-
-    private fun isMessageNewerThanCurrent(currentMessage: Message?, newMessage: Message): Boolean {
-        return if (newMessage.syncStatus == SyncStatus.COMPLETED) {
-            currentMessage?.lastUpdateTime() ?: NEVER.time <= newMessage.lastUpdateTime()
-        } else {
-            currentMessage?.lastLocalUpdateTime() ?: NEVER.time <= newMessage.lastLocalUpdateTime()
-        }
-    }
-
-    internal fun upsertMessages(messages: List<Message>) {
-        val newMessages = parseMessages(messages)
-        updateLastMessageAtByNewMessages(newMessages.values)
-        mutableState._messages.value = newMessages
-    }
-
-    private fun updateLastMessageAtByNewMessages(newMessages: Collection<Message>) {
-        if (newMessages.isEmpty()) {
-            return
-        }
-        val newLastMessageAt =
-            newMessages.mapNotNull { it.createdAt ?: it.createdLocallyAt }.maxOfOrNull(Date::getTime) ?: return
-        mutableState.lastMessageAt.value = when (val currentLastMessageAt = mutableState.lastMessageAt.value) {
-            null -> Date(newLastMessageAt)
-            else -> max(currentLastMessageAt.time, newLastMessageAt).let(::Date)
-        }
-    }
-
-    private fun upsertOldMessages(messages: List<Message>) {
-        mutableState._oldMessages.value = parseMessages(messages)
     }
 
     private fun removeLocalMessage(message: Message) {
@@ -840,10 +742,6 @@ public class ChannelController internal constructor(
         mutableState._typing.value = copy.toMap()
     }
 
-    private fun setHidden(hidden: Boolean) {
-        mutableState._hidden.value = hidden
-    }
-
     internal suspend fun handleEvents(events: List<ChatEvent>) {
         for (event in events) {
             handleEvent(event)
@@ -858,24 +756,24 @@ public class ChannelController internal constructor(
         when (event) {
             is NewMessageEvent -> {
                 upsertEventMessage(event.message)
-                incrementUnreadCountIfNecessary(event.message)
-                setHidden(false)
+                channelLogic.incrementUnreadCountIfNecessary(event.message)
+                channelLogic.setHidden(false)
             }
             is MessageUpdatedEvent -> {
                 event.message.apply {
                     replyTo = mutableState._messages.value[replyMessageId]
                 }.let(::upsertEventMessage)
 
-                setHidden(false)
+                channelLogic.setHidden(false)
             }
             is MessageDeletedEvent -> {
                 upsertEventMessage(event.message)
-                setHidden(false)
+                channelLogic.setHidden(false)
             }
             is NotificationMessageNewEvent -> {
                 upsertEventMessage(event.message)
-                incrementUnreadCountIfNecessary(event.message)
-                setHidden(false)
+                channelLogic.incrementUnreadCountIfNecessary(event.message)
+                channelLogic.setHidden(false)
             }
             is ReactionNewEvent -> {
                 upsertMessage(event.message)
@@ -906,23 +804,23 @@ public class ChannelController internal constructor(
             }
             is UserStartWatchingEvent -> {
                 upsertWatcher(event.user)
-                setWatcherCount(event.watcherCount)
+                channelLogic.setWatcherCount(event.watcherCount)
             }
             is UserStopWatchingEvent -> {
                 deleteWatcher(event.user)
-                setWatcherCount(event.watcherCount)
+                channelLogic.setWatcherCount(event.watcherCount)
             }
             is ChannelUpdatedEvent -> {
-                updateChannelData(event.channel)
+                channelLogic.updateChannelData(event.channel)
             }
             is ChannelUpdatedByUserEvent -> {
-                updateChannelData(event.channel)
+                channelLogic.updateChannelData(event.channel)
             }
             is ChannelHiddenEvent -> {
-                setHidden(true)
+                channelLogic.setHidden(true)
             }
             is ChannelVisibleEvent -> {
-                setHidden(false)
+                channelLogic.setHidden(false)
             }
             is ChannelDeletedEvent -> {
                 removeMessagesBefore(event.createdAt)
@@ -954,11 +852,11 @@ public class ChannelController internal constructor(
             }
             is NotificationInviteAcceptedEvent -> {
                 upsertMember(event.member)
-                updateChannelData(event.channel)
+                channelLogic.updateChannelData(event.channel)
             }
             is NotificationInviteRejectedEvent -> {
                 upsertMember(event.member)
-                updateChannelData(event.channel)
+                channelLogic.updateChannelData(event.channel)
             }
             is NotificationChannelMutesUpdatedEvent -> {
                 mutableState._muted.value = event.me.channelMutes.any { mute ->
@@ -1037,7 +935,7 @@ public class ChannelController internal constructor(
             if (changed) changedMessages.add(message)
         }
         if (changedMessages.isNotEmpty()) {
-            upsertMessages(changedMessages)
+            channelLogic.upsertMessages(changedMessages)
         }
     }
 
@@ -1061,113 +959,13 @@ public class ChannelController internal constructor(
         upsertMembers(listOf(member))
     }
 
-    private fun updateReads(reads: List<ChannelUserRead>) {
-        domainImpl.user.value?.let { currentUser ->
-
-            val currentUserId = currentUser.id
-            val previousUserIdToReadMap = mutableState._reads.value
-            val incomingUserIdToReadMap = reads.associateBy(ChannelUserRead::getUserId).toMutableMap()
-
-            /**
-             * It's possible that the data coming back from the online channel query has a last read date that's
-             * before what we've last pushed to the UI. We want to ignore this, as it will cause an unread state
-             * to show in the channel list.
-             */
-            incomingUserIdToReadMap[currentUserId]?.let { incomingUserRead ->
-
-                // the previous last Read date that is most current
-                val previousLastRead =
-                    mutableState._read.value?.lastRead ?: previousUserIdToReadMap[currentUserId]?.lastRead
-
-                // Use AFTER to determine if the incoming read is more current.
-                // This prevents updates if it's BEFORE or EQUAL TO the previous Read.
-                val shouldUpdateByIncoming = previousLastRead == null || incomingUserRead.lastRead?.inOffsetWith(
-                    previousLastRead,
-                    OFFSET_EVENT_TIME
-                ) == true
-
-                if (shouldUpdateByIncoming) {
-                    mutableState._read.value = incomingUserRead
-                    mutableState._unreadCount.value = incomingUserRead.unreadMessages
-                } else {
-                    // if the previous Read was more current, replace the item in the update map
-                    incomingUserIdToReadMap[currentUserId] = ChannelUserRead(currentUser, previousLastRead)
-                }
-            }
-
-            // always post the newly updated map
-            mutableState._reads.value = (previousUserIdToReadMap + incomingUserIdToReadMap)
-        }
-    }
-
     private fun updateRead(
         read: ChannelUserRead,
     ) {
-        updateReads(listOf(read))
+        channelLogic.updateReads(listOf(read))
     }
 
-    private fun updateDataFromLocalChannel(localChannel: Channel) {
-        localChannel.hidden?.let(::setHidden)
-        mutableState.hideMessagesBefore = localChannel.hiddenMessagesBefore
-        updateDataFromChannel(localChannel)
-    }
-
-    private fun updateOldMessagesFromLocalChannel(localChannel: Channel) {
-        localChannel.hidden?.let(::setHidden)
-        mutableState.hideMessagesBefore = localChannel.hiddenMessagesBefore
-        updateOldMessagesFromChannel(localChannel)
-    }
-
-    internal fun updateDataFromChannel(c: Channel) {
-        // Update all the flow objects based on the channel
-        updateChannelData(c)
-        setWatcherCount(c.watcherCount)
-        updateReads(c.read)
-
-        // there are some edge cases here, this code adds to the members, watchers and messages
-        // this means that if the offline sync went out of sync things go wrong
-        setMembers(c.members)
-        setWatchers(c.watchers)
-        upsertMessages(c.messages)
-        mutableState.lastMessageAt.value = c.lastMessageAt
-    }
-
-    private fun updateOldMessagesFromChannel(c: Channel) {
-        // Update all the flow objects based on the channel
-        updateChannelData(c)
-        setWatcherCount(c.watcherCount)
-        updateReads(c.read)
-
-        // there are some edge cases here, this code adds to the members, watchers and messages
-        // this means that if the offline sync went out of sync things go wrong
-        setMembers(c.members)
-        setWatchers(c.watchers)
-        upsertOldMessages(c.messages)
-    }
-
-    private fun setMembers(members: List<Member>) {
-        mutableState._members.value = (mutableState._members.value + members.associateBy(Member::getUserId))
-    }
-
-    private fun updateChannelData(channel: Channel) {
-        mutableState._channelData.value = (ChannelData(channel))
-    }
-
-    private fun setWatchers(watchers: List<User>) {
-        mutableState._watchers.value = (mutableState._watchers.value + watchers.associateBy { it.id })
-    }
-
-    private fun incrementUnreadCountIfNecessary(message: Message) {
-        val currentUserId = domainImpl.user.value?.id
-        if (currentUserId?.let(message::shouldIncrementUnreadCount) == true) {
-            val newUnreadCount = mutableState._unreadCount.value + 1
-            mutableState._unreadCount.value = newUnreadCount
-            mutableState._read.value = mutableState._read.value?.copy(unreadMessages = newUnreadCount)
-            mutableState._reads.value = mutableState._reads.value.apply {
-                this[currentUserId]?.unreadMessages = newUnreadCount
-            }
-        }
-    }
+    internal fun updateDataFromChannel(c: Channel) = channelLogic.updateDataFromChannel(c)
 
     internal suspend fun editMessage(message: Message): Result<Message> {
         // TODO: should we rename edit message into update message to be similar to llc?
@@ -1289,7 +1087,7 @@ public class ChannelController internal constructor(
         }
         if (result.isSuccess) {
             val newMessages = result.data()
-            upsertMessages(newMessages)
+            channelLogic.upsertMessages(newMessages)
             // Note that we don't handle offline storage for threads at the moment.
         }
         return result
@@ -1318,22 +1116,6 @@ public class ChannelController internal constructor(
     }
 
     internal fun cancelJobs() = messageSendingService.cancelJobs()
-
-    private fun Message.lastUpdateTime(): Long = listOfNotNull(
-        createdAt,
-        updatedAt,
-        deletedAt,
-    ).map { it.time }
-        .maxOrNull()
-        ?: NEVER.time
-
-    private fun Message.lastLocalUpdateTime(): Long = listOfNotNull(
-        createdLocallyAt,
-        updatedLocallyAt,
-        deletedAt,
-    ).map { it.time }
-        .maxOrNull()
-        ?: NEVER.time
 
     public sealed class MessagesState {
         /** The ChannelController is initialized but no query is currently running.
@@ -1384,6 +1166,5 @@ public class ChannelController internal constructor(
         private const val KEY_MESSAGE_ACTION = "image_action"
         private const val MESSAGE_ACTION_SHUFFLE = "shuffle"
         private const val MESSAGE_ACTION_SEND = "send"
-        private const val OFFSET_EVENT_TIME = 5L
     }
 }
