+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/experimental/channel/logic/ChannelLogic.kt
@@ -1,0 +1,279 @@
+package io.getstream.chat.android.offline.experimental.channel.logic
+
+import io.getstream.chat.android.client.api.models.QueryChannelRequest
+import io.getstream.chat.android.client.errors.ChatError
+import io.getstream.chat.android.client.experimental.plugin.listeners.QueryChannelListener
+import io.getstream.chat.android.client.logger.ChatLogger
+import io.getstream.chat.android.client.models.Channel
+import io.getstream.chat.android.client.models.ChannelUserRead
+import io.getstream.chat.android.client.models.Member
+import io.getstream.chat.android.client.models.Message
+import io.getstream.chat.android.client.models.User
+import io.getstream.chat.android.client.utils.Result
+import io.getstream.chat.android.client.utils.SyncStatus
+import io.getstream.chat.android.client.utils.onError
+import io.getstream.chat.android.client.utils.onSuccess
+import io.getstream.chat.android.client.utils.onSuccessSuspend
+import io.getstream.chat.android.core.ExperimentalStreamChatApi
+import io.getstream.chat.android.offline.ChatDomainImpl
+import io.getstream.chat.android.offline.channel.ChannelData
+import io.getstream.chat.android.offline.experimental.channel.state.ChannelMutableState
+import io.getstream.chat.android.offline.extensions.inOffsetWith
+import io.getstream.chat.android.offline.extensions.isPermanent
+import io.getstream.chat.android.offline.message.NEVER
+import io.getstream.chat.android.offline.message.attachment.AttachmentUrlValidator
+import io.getstream.chat.android.offline.message.shouldIncrementUnreadCount
+import io.getstream.chat.android.offline.model.ChannelConfig
+import java.util.Date
+import kotlin.math.max
+
+@ExperimentalStreamChatApi
+internal class ChannelLogic(
+    private val mutableState: ChannelMutableState,
+    private val chatDomainImpl: ChatDomainImpl,
+    private val attachmentUrlValidator: AttachmentUrlValidator = AttachmentUrlValidator(),
+) : QueryChannelListener {
+
+    private val logger = ChatLogger.get("Query channel request")
+
+    internal fun loadingStateByRequest(request: QueryChannelRequest) = when {
+        request.isFilteringNewerMessages() -> mutableState._loadingNewerMessages
+        request.filteringOlderMessages() -> mutableState._loadingOlderMessages
+        else -> mutableState._loading
+    }
+
+    override suspend fun onQueryChannelPrecondition(
+        channelType: String,
+        channelId: String,
+        request: QueryChannelRequest,
+    ): Result<Unit> {
+        val loader = loadingStateByRequest(request)
+        return if (loader.value) {
+            logger.logI("Another request to load messages is in progress. Ignoring this request.")
+            Result.error(ChatError("Another request to load messages is in progress. Ignoring this request."))
+        } else {
+            Result.success(Unit)
+        }
+    }
+
+    override suspend fun onQueryChannelRequest(channelType: String, channelId: String, request: QueryChannelRequest) {
+        runChannelQueryOffline(request)
+    }
+
+    override suspend fun onQueryChannelResult(
+        result: Result<Channel>,
+        channelType: String,
+        channelId: String,
+        request: QueryChannelRequest,
+    ) {
+        result.onSuccessSuspend { channel ->
+            // first thing here needs to be updating configs otherwise we have a race with receiving events
+            chatDomainImpl.repos.insertChannelConfig(ChannelConfig(channel.type, channel.config))
+            chatDomainImpl.storeStateForChannel(channel)
+        }
+            .onSuccess { channel ->
+                mutableState.recoveryNeeded = false
+                if (request.messagesLimit() > channel.messages.size) {
+                    if (request.isFilteringNewerMessages()) {
+                        mutableState._endOfNewerMessages.value = true
+                    } else {
+                        mutableState._endOfOlderMessages.value = true
+                    }
+                }
+                updateDataFromChannel(channel)
+                loadingStateByRequest(request).value = false
+            }
+            .onError { error ->
+                if (error.isPermanent()) {
+                    logger.logW("Permanent failure calling channel.watch for channel ${mutableState.cid}, with error $error")
+                } else {
+                    logger.logW("Temporary failure calling channel.watch for channel ${mutableState.cid}. Marking the channel as needing recovery. Error was $error")
+                    mutableState.recoveryNeeded = true
+                }
+                chatDomainImpl.addError(error)
+            }
+    }
+
+    internal suspend fun runChannelQueryOffline(request: QueryChannelRequest): Channel? {
+        val loader = loadingStateByRequest(request)
+        loader.value = true
+        return chatDomainImpl.selectAndEnrichChannel(mutableState.cid, request)?.also { channel ->
+            logger.logI("Loaded channel ${channel.cid} from offline storage with ${channel.messages.size} messages")
+            if (request.filteringOlderMessages()) {
+                updateOldMessagesFromLocalChannel(channel)
+            } else {
+                updateDataFromLocalChannel(channel)
+            }
+            loader.value = false
+        }
+    }
+
+    private fun updateDataFromLocalChannel(localChannel: Channel) {
+        localChannel.hidden?.let(::setHidden)
+        mutableState.hideMessagesBefore = localChannel.hiddenMessagesBefore
+        updateDataFromChannel(localChannel)
+    }
+
+    private fun updateOldMessagesFromLocalChannel(localChannel: Channel) {
+        localChannel.hidden?.let(::setHidden)
+        mutableState.hideMessagesBefore = localChannel.hiddenMessagesBefore
+        updateOldMessagesFromChannel(localChannel)
+    }
+
+    internal fun setHidden(hidden: Boolean) {
+        mutableState._hidden.value = hidden
+    }
+
+    private fun updateOldMessagesFromChannel(c: Channel) {
+        // Update all the flow objects based on the channel
+        updateChannelData(c)
+        setWatcherCount(c.watcherCount)
+        updateReads(c.read)
+
+        // there are some edge cases here, this code adds to the members, watchers and messages
+        // this means that if the offline sync went out of sync things go wrong
+        setMembers(c.members)
+        setWatchers(c.watchers)
+        upsertOldMessages(c.messages)
+    }
+
+    private fun upsertOldMessages(messages: List<Message>) {
+        mutableState._oldMessages.value = parseMessages(messages)
+    }
+
+    internal fun updateDataFromChannel(c: Channel) {
+        // Update all the flow objects based on the channel
+        updateChannelData(c)
+        setWatcherCount(c.watcherCount)
+        updateReads(c.read)
+
+        // there are some edge cases here, this code adds to the members, watchers and messages
+        // this means that if the offline sync went out of sync things go wrong
+        setMembers(c.members)
+        setWatchers(c.watchers)
+        upsertMessages(c.messages)
+        mutableState.lastMessageAt.value = c.lastMessageAt
+    }
+
+    internal fun upsertMessages(messages: List<Message>) {
+        val newMessages = parseMessages(messages)
+        updateLastMessageAtByNewMessages(newMessages.values)
+        mutableState._messages.value = newMessages
+    }
+
+    internal fun setWatcherCount(watcherCount: Int) {
+        if (watcherCount != mutableState._watcherCount.value) {
+            mutableState._watcherCount.value = watcherCount
+        }
+    }
+
+    private fun setMembers(members: List<Member>) {
+        mutableState._members.value = (mutableState._members.value + members.associateBy(Member::getUserId))
+    }
+
+    internal fun updateChannelData(channel: Channel) {
+        mutableState._channelData.value = (ChannelData(channel))
+    }
+
+    private fun setWatchers(watchers: List<User>) {
+        mutableState._watchers.value = (mutableState._watchers.value + watchers.associateBy { it.id })
+    }
+
+    internal fun incrementUnreadCountIfNecessary(message: Message) {
+        val currentUserId = chatDomainImpl.user.value?.id
+        if (currentUserId?.let(message::shouldIncrementUnreadCount) == true) {
+            val newUnreadCount = mutableState._unreadCount.value + 1
+            mutableState._unreadCount.value = newUnreadCount
+            mutableState._read.value = mutableState._read.value?.copy(unreadMessages = newUnreadCount)
+            mutableState._reads.value = mutableState._reads.value.apply {
+                this[currentUserId]?.unreadMessages = newUnreadCount
+            }
+        }
+    }
+
+    internal fun updateReads(reads: List<ChannelUserRead>) {
+        chatDomainImpl.user.value?.let { currentUser ->
+
+            val currentUserId = currentUser.id
+            val previousUserIdToReadMap = mutableState._reads.value
+            val incomingUserIdToReadMap = reads.associateBy(ChannelUserRead::getUserId).toMutableMap()
+
+            /**
+             * It's possible that the data coming back from the online channel query has a last read date that's
+             * before what we've last pushed to the UI. We want to ignore this, as it will cause an unread state
+             * to show in the channel list.
+             */
+            incomingUserIdToReadMap[currentUserId]?.let { incomingUserRead ->
+
+                // the previous last Read date that is most current
+                val previousLastRead =
+                    mutableState._read.value?.lastRead ?: previousUserIdToReadMap[currentUserId]?.lastRead
+
+                // Use AFTER to determine if the incoming read is more current.
+                // This prevents updates if it's BEFORE or EQUAL TO the previous Read.
+                val shouldUpdateByIncoming = previousLastRead == null || incomingUserRead.lastRead?.inOffsetWith(
+                    previousLastRead,
+                    OFFSET_EVENT_TIME
+                ) == true
+
+                if (shouldUpdateByIncoming) {
+                    mutableState._read.value = incomingUserRead
+                    mutableState._unreadCount.value = incomingUserRead.unreadMessages
+                } else {
+                    // if the previous Read was more current, replace the item in the update map
+                    incomingUserIdToReadMap[currentUserId] = ChannelUserRead(currentUser, previousLastRead)
+                }
+            }
+
+            // always post the newly updated map
+            mutableState._reads.value = (previousUserIdToReadMap + incomingUserIdToReadMap)
+        }
+    }
+
+    private fun parseMessages(messages: List<Message>): Map<String, Message> {
+        val currentMessages = mutableState._messages.value
+        return currentMessages + attachmentUrlValidator.updateValidAttachmentsUrl(messages, currentMessages)
+            .filter { newMessage -> isMessageNewerThanCurrent(currentMessages[newMessage.id], newMessage) }
+            .associateBy(Message::id)
+    }
+
+    private fun isMessageNewerThanCurrent(currentMessage: Message?, newMessage: Message): Boolean {
+        return if (newMessage.syncStatus == SyncStatus.COMPLETED) {
+            currentMessage?.lastUpdateTime() ?: NEVER.time <= newMessage.lastUpdateTime()
+        } else {
+            currentMessage?.lastLocalUpdateTime() ?: NEVER.time <= newMessage.lastLocalUpdateTime()
+        }
+    }
+
+    private fun updateLastMessageAtByNewMessages(newMessages: Collection<Message>) {
+        if (newMessages.isEmpty()) {
+            return
+        }
+        val newLastMessageAt =
+            newMessages.mapNotNull { it.createdAt ?: it.createdLocallyAt }.maxOfOrNull(Date::getTime) ?: return
+        mutableState.lastMessageAt.value = when (val currentLastMessageAt = mutableState.lastMessageAt.value) {
+            null -> Date(newLastMessageAt)
+            else -> max(currentLastMessageAt.time, newLastMessageAt).let(::Date)
+        }
+    }
+
+    private fun Message.lastUpdateTime(): Long = listOfNotNull(
+        createdAt,
+        updatedAt,
+        deletedAt,
+    ).map { it.time }
+        .maxOrNull()
+        ?: NEVER.time
+
+    private fun Message.lastLocalUpdateTime(): Long = listOfNotNull(
+        createdLocallyAt,
+        updatedLocallyAt,
+        deletedAt,
+    ).map { it.time }
+        .maxOrNull()
+        ?: NEVER.time
+
+    private companion object {
+        private const val OFFSET_EVENT_TIME = 5L
+    }
+}
