--- stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/experimental/querychannels/logic/QueryChannelLogic.kt
@@ -1,56 +1,0 @@
-package io.getstream.chat.android.offline.experimental.querychannels.logic
-
-import io.getstream.chat.android.client.api.models.QueryChannelRequest
-import io.getstream.chat.android.client.errors.ChatError
-import io.getstream.chat.android.client.experimental.plugin.listeners.QueryChannelListener
-import io.getstream.chat.android.client.logger.ChatLogger
-import io.getstream.chat.android.client.models.Channel
-import io.getstream.chat.android.client.utils.Result
-import io.getstream.chat.android.client.utils.onSuccessSuspend
-import io.getstream.chat.android.core.ExperimentalStreamChatApi
-import io.getstream.chat.android.offline.ChatDomainImpl
-import io.getstream.chat.android.offline.experimental.channel.state.ChannelMutableState
-
-@ExperimentalStreamChatApi
-internal class QueryChannelLogic(
-    private val mutableState: ChannelMutableState,
-    private val chatDomainImpl: ChatDomainImpl,
-) : QueryChannelListener {
-
-    private val logger = ChatLogger.get("Query channel request")
-
-    internal fun loadingStateByRequest(request: QueryChannelRequest) = when {
-        request.isFilteringNewerMessages() -> mutableState._loadingNewerMessages
-        request.filteringOlderMessages() -> mutableState._loadingOlderMessages
-        else -> mutableState._loading
-    }
-
-    override suspend fun onQueryChannelPrecondition(
-        channelType: String,
-        channelId: String,
-        request: QueryChannelRequest,
-    ): Result<Unit> {
-        val loader = loadingStateByRequest(request)
-        return if (loader.value) {
-            logger.logI("Another request to load messages is in progress. Ignoring this request.")
-            Result.error(ChatError("Another request to load messages is in progress. Ignoring this request."))
-        } else {
-            Result.success(Unit)
-        }
-    }
-
-    override suspend fun onQueryChannelRequest(channelType: String, channelId: String, request: QueryChannelRequest) {
-        super.onQueryChannelRequest(channelType, channelId, request)
-    }
-
-    override suspend fun onQueryChannelResult(
-        result: Result<Channel>,
-        channelType: String,
-        channelId: String,
-        request: QueryChannelRequest,
-    ) {
-        result.onSuccessSuspend {
-            chatDomainImpl.repos.insertChannel(it)
-        }
-    }
-}
